CVSL-3438 Time-Served-Retain-selection-and-free-text-if-NOMIS-selected

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/createandvaryalicenceapi/entity/NomisTimeServedLicence.kt
@@ -15,7 +15,7 @@ class NomisTimeServedLicence(
   val nomsId: String,
 
   @Column(name = "booking_id", nullable = false)
-  val bookingId: Int,
+  val bookingId: Long,
 
   @Column(name = "reason", columnDefinition = "TEXT", nullable = false)
   var reason: String,
